requirements.txt
anthropic
cryptography
google-generativeai
numpy
openai
python-dotenv
flask
flask-sqlalchemy
flask-login
flask-wtf
email-validator